requirements.txt
fastapi==0.85.0
gunicorn==20.1.0
uvicorn==0.22.0
requests==2.28.1
requests-oauthlib==1.3.1
google-cloud-run
google-cloud-compute
grpc-google-iam-v1
